// src/redux/aidiagnosis/types.ts
export enum AiDiagnosisActionTypes {
  FETCH_SPINDLE_BY_ID_REQUEST = 'FETCH_SPINDLE_BY_ID_REQUEST',
  FETCH_SPINDLE_BY_ID_SUCCESS = 'FETCH_SPINDLE_BY_ID_SUCCESS',
  FETCH_SPINDLE_BY_ID_FAILURE = 'FETCH_SPINDLE_BY_ID_FAILURE',

  FETCH_CLASSIFICATION_REQUEST = 'FETCH_CLASSIFICATION_REQUEST',
  FETCH_CLASSIFICATION_SUCCESS = 'FETCH_CLASSIFICATION_SUCCESS',
  FETCH_CLASSIFICATION_FAILURE = 'FETCH_CLASSIFICATION_FAILURE',

  FETCH_DIAGNOSIS_REQUEST = 'FETCH_DIAGNOSIS_REQUEST',
  FETCH_DIAGNOSIS_SUCCESS = 'FETCH_DIAGNOSIS_SUCCESS',
  FETCH_DIAGNOSIS_FAILURE = 'FETCH_DIAGNOSIS_FAILURE',
}